Remove unnecessary SandboxTestCase base class.
The contents of SandboxTestCase were trivial at this point, so keeping the
class around yields more confusion than benefit.  In particular, the
presence of a "testRoot" field inherited from a parent test class prevents
tweaking the file system on which the test root lives, which is something
I want to do in an upcoming test.

RELNOTES: None.
PiperOrigin-RevId: 230761854