Only return SkylarkValue actions in target.actions
This fixes a Starlark conversion type exception when attempting to access actions with no Starlark API.

This marks a tiny behavioral change in that aspects will thus return fewer actions, but this should not be treated as an incompatible change, as any attempt to invoke accessors of these now-dropped actions would result in a RuntimeException.

RELNOTES: None.
PiperOrigin-RevId: 230766076