Move depset order validation from PyCommon to PyStructUtils
The fact that the field is a depset is already an invariant of the provider, so we may as well include the order in that requirement. This is not a breaking change because it's currently enforced by PyCommon.

Moving the validation out of PyCommon will make it easier to refactor the "get this field from that dep" helpers out of PyCommon and into a shared PyProviderUtils helper.

Work toward #7010.

RELNOTES: None
PiperOrigin-RevId: 230780935