Add helper for abstracting over Python provider representation and defaults
This creates PyProviderUtils, a collection of helpers that take in a TransitiveInfoCollection and return fields obtained from the Python provider. Currently this just dispatches to PyStructUtils to retrieve from the legacy "py" struct provider. In a follow-up CL, we'll add a modern PyInfo type, and PyProviderUtils will access either one as appropriate.

PyProviderUtils also takes over the default logic from PyCommon, for when no Python provider information is available.

Work toward #7010.

RELNOTES: None
PiperOrigin-RevId: 230791581