Add the ability to sort manifests by root-relative path instead of full paths.
Currently, mergee manifests are listed in path order. However, this may result
in erroneous conflicts. Dependency order, with the manifests of dependencies
listed after the manifests of the targets which depend on them, allows for
proper overriding.

However, due to how resources are passed along the android rules, dependency
order would be incorrect without making changes to that order, which could
break apps that are relying on that order. And it would require significant
work to allow for the order change to be optional.

(Specifically, the nested set needs to be link order, not naive link order,
and the nested set's items need to be put in with nested sets representing
direct dependencies first, and nested sets representing transitive dependencies
second.)

As a result, this change instead offers the ability to sort by root-relative
paths, which will drop the configuration directory from the path for comparison
purposes.

A dependency option is available; however, the order of that option is
incorrect.

RELNOTES: None.
PiperOrigin-RevId: 230797609