Add an aspect to @bazel_tools for debugging bad Python dependencies
We need this aspect because under the new Python version semantics, incompatibilities in srcs_version are caught by py_binary rather than at the py_library where the mismatch occurred. The workflow is for users to see the validation error and follow a URL to documentation (to come in a later CL) instructing them to run this aspect to locate the bad dependencies.

The aspect propagates to deps and uses both srcs_version and the py provider fields to try to determine the top-most targets that require Python 2 or Python 3. It's possible this idiom of "find the topmost dependencies satisfying property X" is reusable and could be factored into Skylib (perhaps once this code is moved out of core Bazel and into rules_python).

Incidentally removed python_version_srcs filegroup in favor of just using srcs.

Work towards #6583 and #1446.

RELNOTES: None
PiperOrigin-RevId: 230813240