Handling exceptions from test runner for graceful termination
If an unhandled exception occurs during the execution of a test suite, it will cause the TestRunner process to hang, and only be terminated after a timeout from the executing strategy.
This PR wraps the actual test execution with a try/catch, allowing the test runner to report the error and gracefully terminate. This allows the exception text be written to `test.log`, and the executing strategy will create a `test.xml` out of it, so the error is correctly reported in the tools (such as Intellij)

Note: added the snippet to both `BazelTestRunner` and the `ExperimentalTestRunner`, but not sure if the latter is needed.

Closes #7192.

PiperOrigin-RevId: 230944912