Add --check_third_party_targets_have_licenses.
When on, all rules and files declared with exports_file()
under //third_party must declare licenses().

This is the flag's default, and also status quo semantics.
The point of this flag is to let users turn off that
checking when they don't want to treat //third_party as
special.

Fixes #3639