Prevent make_hashed_objlist from writing into undeclared locations.
The make_hashed_objlist script was creating symlinks to object files in
the same directory where those object files lived.  This is a problem
because there is no guarantee that we can write to those directories as
demonstrated by the stricter sandboxfs-sandboxing.

To fix this, make this script create its links to a given temporary
directory and fix the callers to create such directory.

Note that SandboxfsSandboxedSpawn currently contains code to cope with
this by creating all hierarchies that correspond to input files.  This
seems like a hack and is costly, hence why I'd like to remove it.

RELNOTES: None.
PiperOrigin-RevId: 231206610