Fold BazelCrashUtils.java into BugReport.java. There's no need to have both of these.
This was very tedious and a lot of care had to be done in order to avoid build-time cycles -- this is mostly due to the combination of (a) the serialization codebase's usage of BugReport and (b) the proliferation of '@AutoCodec' annotations.

Also, while I'm here, slightly improve readability of BugReport.java.

RELNOTES: None
PiperOrigin-RevId: 231241644