Do not create writable directory trees for input files.
The sandboxfs sandboxed spawn was creating writable file hierarchies
for all input files.  This should not be necessary if actions are
well-behaved.  Therefore, remove this for stricter sandboxing and to
improve sandboxfs-based performance by avoiding a significant number
of directory operations per action.

More specifically, this code was put in place to workaround an issue
with linking on macOS where a helper tool (make_hashed_objlist.py)
wanted to create symlinks in the directories where inputs lived.
I don't know why I concluded that the sandboxing code was wrong
instead of thinking that the tooling had to be fixed.  I have now
fixed the offending tools in unknown commit so we can do this
simplification.

Addresses part of https://github.com/bazelbuild/bazel/issues/7273.

Tested: Built a relatively large iOS app with this change and the
only problem I found is the aforementioned one in
make_hashed_objlist.py.  It is possible that this strictness change
breaks other rules but the success of this build makes me relatively
confident that problems are not widespread.

RELNOTES: None.
PiperOrigin-RevId: 231242503