Do not use NestedSetVisitor's custom NestedSet traversal logic when replaying events and postables. Instead, just walk the NestedSet.
NestedSetVisitor allows online visitation of NestedSets and promises to re-use
results from visited subsets so that using incrementally running over every
entry of a large NestedSet structure should be efficient. Using
ConcurrentHashMap makes this thread safe.

However, this has a lot of overhead. A ConcurrentHashMap entry is stored for
every node (leaf or no) in the NestedSet. In huge NestedSet structures that have
few actual entries, as it is the case for the Events and Postables it is
currently used for, this is especially bad. When replaying events, this is
usually only done for a limited number of leaf nodes and thus using the
regular NestedSet visitation is more efficient.

RELNOTES: None.
PiperOrigin-RevId: 231384665