Warn if moving outputs out of the sandbox requires doing file copies.
Extend the FileSystemUtils.moveFile function to return information on how
the move actually happened, and then use this information in the
moveOutputs() function to warn if a copy happened.  Having to copy outputs
out of the sandbox could be very costly.

For now just keep the warning in the server log so we can troubleshoot
strange slowdowns.  Would be nice to surface this to the user as part of
regular Blaze output but we don't have access to the reporter from this
code and plumbing it through seems hard/overkill.

RELNOTES: None.
PiperOrigin-RevId: 231437825