Always set --no-canonical-prefixes if we can
Some configurations of gcc (I suspect --without-sysroot is the key
flag, but it's tricky to prove) will still resolve symlinks in the
compiler path even when -fno-canonical-system-headers is used. This
breaks remote execution in the presence of symlinks.

There's not much we can do other than include both flags - it's hard
to tell what will be correct when executing remotely. However, there
appears to be no real harm in adding --no-canonical-prefixes if it
passes the test in _is_linker_option_supported: prefix
canonicalisation is only required when gcc is invoked through a
symlink that does not also include the rest of its files, and the test
command will fail completely when this occurs.