Add an optional output_source_jar parameter to java_common.compile
If provided, the artifact is used for the generated source jar output.
Otherwise the output source jar is written to `{output_jar}-src.jar`.
This parameter is useful for rules that want to customize how the
artifact is declared (e.g. with a different path, or as a declared
output of the owning rule).

RELNOTES: Add an optional output_source_jar parameter to java_common.compile
PiperOrigin-RevId: 231845401