Always set a trap for SIGTERM in test-setup.sh
If we don't set a trap here, then bash ignores the signal, and the test
process also does not receive the signal, so the test runner has no
chance of writing a test.xml output.

However, the behavior of trap forwarding the signal to the subprocess is
not at all documented in the bash documentation, and also inconsistent
with the behavior reported in #7119.

There is a similar problem in the Java stub template reported in #6338.

This may or may not be progress on #4608.

PiperOrigin-RevId: 232035930