Simplify WorkspaceStatusAction handling
It was using multiple different mechanisms to inject runtime data.
Instead, use the ActionExecutionContext for all variable data; the only
thing that needs to be set for the actions at construction time are the
stable and volatile output artifacts.

This removes the need for custom serialization handling, as well as
fixing existing bugs in Bazel around incorrect runtime injection.

Furthermore, change the code to use an Environment mechanism for both
the workspace status action and its dummy equivalent (which doesn't
exist in Bazel right now). This makes the *WorkspaceStatusModule classes
completely stateless, which is also a step towards concurrent execution
of commands in the same server.

PiperOrigin-RevId: 232065580