Avoid deletion of repository directory in repository_ctx
On Windows, it is impossible to delete a directory while this directory
is the working directory of a running process. In _clone_or_update,
however, BAZEL_SH is called from the repository_ctx, giving it the
repository directory as working directory. The cleaning of the
directory is then performed by deleting and re-creating that very
directory.

Cygwin apparently does some workarounds to make this possible, but
other shells (e. g. busybox-w32) don't. This commit changes the rule to
cd out of the repository working directory first to avoid locking.

Fixes #7274.