Default --experimental_split_xml_generation to true
This changes the test action to generate the xml file as a separate step
if the test itself did not generate an xml file in the first place. The
reason for this change is a race condition in Posix-style operating
systems when piping output to stdout and to a file simultaneously.

Fixes #4608.

RELNOTES: For tests that do not generate a test.xml, Bazel now uses a separate action to generate one; this results in minor differences in the generated test.xml, and makes the generation more reliable overall.
PiperOrigin-RevId: 232351927