Invert static_link_test_srcs feature to dynamic_link_test_srcs.
The way the logic of the feature works at the moment, we can disable it globally but we can't enable it for specific rules because the global setting overrides the rule setting. We need to invert the feature to be able to enable it globally, but disable it locally for targets that don't work with the feature. Push back the retirement timeline for the experimental flag as well.

RELNOTES: None.
PiperOrigin-RevId: 232438035