Remove the originating rule and attribute from the "inconsistent aspect order" error message.
This is unfortunate, but necessary for separating the configuration and aspect computation of a dependency into a Skyframe function that's separate from the evaluation of its parent.

It's conceivable that this could be restored by emitting the message when the unfiltered aspect path is computed, but let's do that in a followup change if necessary. I'm not convinced that we need this at all...

RELNOTES: None.
PiperOrigin-RevId: 232456350