experimental UI: change "no action" default message
After finishing analysis phase, bazel might not start executing
any action immediately---when traversing the action graph, a lot
of actions might be cached, but that checking if an action is cached
still takes time. In this case, the experimental UI shows "no action"
as current activity, which is technically correct, but seems to confuse
some users. Therefore change the message to a hopefully less confusing
one. While there, also make the no-activity message depend on whether
loading has not yet started or analysis is already finished.

Change-Id: I3f09f8f8f2aa697177819d6534e3ec3f2c36b181
PiperOrigin-RevId: 232466726