Pass the correct import path to protoc when using strict dependency checks or exports
Currently, when strict dependency checks are enabled Bazel passes the full path of each directly dependent proto file to protoc via the `--direct_dependencies` flag. Similarly, when proto exports are used the full path of each export is passed to protoc via the `--allowed_public_imports` flag.

With the introduction of the `import_path` and `strip_import_path` proto_library attributes, it is no longer true that the full path of a .proto file is always its import path. Meaning that when the import path is modified, Bazel passes the incorrect import path to the `direct_dependencies` and `allowed_public_imports` protoc flags.

This issue is fixed by having `ProtoInfo` keep track of a pair of each .proto file and its import path. That way whenever the import path has been modified it can be used as the protoc command line argument rather than the .proto's full path.

Closes #7280.

PiperOrigin-RevId: 232661815