ExecutionProgressReceiver: keep the invariant that enqueued is a superset of completed
While the order of invocation of the methods should normally keep the
invariant that the  set of enqueued actions is a superset of the set
of completed actions this does not seem to be always the case, as recent
invocations show. Therefore, ensure that every single method keeps this
invariant.

Change-Id: Ie81eb9a0e834e88379c1436424c05023d8c439f3
PiperOrigin-RevId: 232669752