Have repository functions delete files before writing to them
It is common for underlying files in repositories to be populated with
symlinks, and then have BUILD files overlaid. If we attempt to use
repository_ctx.template or repository_ctx.file on a filename which has
already been created as a symlink, bazel will currently attempt to
write through this symlink, and often fail because the underlying file
is read-only (or worse, succeed and mutate something outside the
repository directory).

We solve this by deleting any pre-existing files before writing new
ones.

Closes #7317.

PiperOrigin-RevId: 232673159