Report the "toolchain cannot be resolved" error in DependencyResolver.
This allows us to remove the special-casing of toolchain dependencies.

The alternative would be to inject the toolchain after the DependencyResolver#dependentNodeMap() call. This would let us do away with the (arguably ugly) toolchainLabels argument to the above method and keep the error message as it is.

However, it would also exempt toolchains from being subject to the configuration transition associated with their rule class (or else that would have to be added in as a special case). They *are* exempt at HEAD, so it wouldn't be a regression, but it seems that that is only an accidental thing and not by design. (not that it matters much because currently all toolchains are in the host configuration and that doesn't have any outgoing transition edges..)

RELNOTES: None.
PiperOrigin-RevId: 232680671