Automated rollback of commit e2e235359ccdc4b6a122586b9d7c99abbddd65f5.
*** Reason for rollback ***

Breaks //devtools/blaze/integration:signed_target_summary_test .

*** Original change description ***

Fix blatant logic bug in FileFunction that causes incorrect FileValues on incremental evaluations.

For a path like "<prefix>/b", we should be using FileStateValue(r/b), where "r" is the fully resolved path of "<prefix>", when we construct the FileValue. This is important in the case where "<prefix>" has a different resolved path (due to ancestor directory symlinks). This was the intent of the FileFunction + FileStateValue algorithm, but the code's structure hid the bug. While I'm here, also sli...

***

RELNOTES: None.
PiperOrigin-RevId: 232836243