C++: Do not look at nocopts in usePicForBinaries
This required looking at the rule context and we want to stop doing that from behind the C++ Starlark API. This was only used by one target in the depot that is broken anyway. Temporarily in case this breaks anyone in Bazel, rather than using an incompatible flag, we can fix anyone broken by this using a rule feature. If no one complains, this replacement will be removed in a later release.

#4570

RELNOTES:none
PiperOrigin-RevId: 232848408