Fix issue with large stdout/stderr payloads being sent over BEP atomically. While there is buffering/flushing logic built into the SynchronizedOutputStream, it did not split large atomic writes.
Now the SOS will split large payloads, requiring that the BuildEventStreamer handle multiple chunks gracefully.

To maintain ordering, we place messages with stdout first, then message with stderr. For backwards compatibility with unit tests and to optimize the number of messages sent, the final stdout message may include the start of the stderr payload.

RELNOTES: None
PiperOrigin-RevId: 232956532