Retry progressive gRPC blob downloads
A gRPC downloadBlob can be distributed over multiple sequential
ByteStream::read requests, all of which will fail prior to ultimate
completion or failure of the download, represented by a single
ListenableFuture. To provide this abstraction,
GrpcRemoteCache::downloadBlob supplies the requests with an offset that
transits committed size between each request, a Retrier.Backoff that is
reset each time partial content is received, and an optional hash
supplier attached to a digesting filter bound to the lifetime of the
download. Each request retains the count of bytes that has been
committed to the output stream since it began.

When a request error is encountered during a gRPC blob download, the
client proceeds as follows:

  If the error is not retriable as determined by the GrpcActionCache's
retrier, the blob download future fails.
  If the error is retriable and any positive size of partial content has
been committed to the output stream since the request began, a new
request is initiated with a read from the current request offset plus
the committed size.
  If the error is retriable and no partial content has been committed to
the output stream, a delay is acquired from its backoff. If the delay is
negative, the download fails immediately with the error. Otherwise a new
request is initiated with a read from the current request offset.

All retry and grpc Context logic has been removed from the
AbstractRemoteActionCache, making the individual ActionCache/store
implementations responsible for their own retry mechanisms.