Try extra hard to not leave a sandboxfs behind on shutdown.
Make every effort we can to terminate the sandboxfs process on Bazel
shutdown, both during clean exits and during crashes.

Otherwise, if the previous build was run with --sandbox_debug or did
not correctly complete for any reason, we'd leave the mount point
busy... which is very annoying because that'd cause the next build to
fail.  This would happen even if the next build didn't use sandboxfs
because the sandboxing code would encounter the sandboxfs mount point,
try to delete it, and fail.

This should fix https://github.com/bazelbuild/bazel/issues/5271.

RELNOTES: None.
PiperOrigin-RevId: 233595796