Progress reporting: also note exceptional end of fetching
When a repository is fetched, the "experimental UI" also reports
about the progress of this process. It does so by assuming that
the latest reported state remains until something new is reported.
Now, a `fetch` may also terminate by an exception; ensure that
the end of the fetching is reported in this case as well.

Change-Id: I4a87a0bf400e01e954f44a8b87b72e7bd37b3312
PiperOrigin-RevId: 233610981