Make Python rules properly read the `imports` field of PyInfo
This makes the built-in Python rules actually accept import paths passed in through the `imports` field of PyInfo (or the deprecated legacy "py" provider). Previously, this field was treated as write-only by the rules.

This is also a code simplification: We get to delete PythonImportsProvider and simplify the dataflow around the imports field of PyCommon.

Fixes #7054.

RELNOTES: None
PiperOrigin-RevId: 233635004