Improve reporting of return vales from repository rules
Repository rules may return a value, indicating a version to make
them reproducible in a canonical way (e.g., by adding commit ids, or
the information needed to allow shallow cloning). This information is
recorded in a resolved file, if one is used; but even if none is
used, this information is useful for the user who might want to
manually change the repository definition in their WORKSPACE file.
However, the current wording ("Rule ... modified arguments ...")
did not convey that use clearly; change the wording to help putting
that information into context.

Fixes #7389.

Change-Id: I9f1044760126f4b7ec4a3d6bf35eaa3568522bb8
PiperOrigin-RevId: 233917633