SkylarkRepositoryContext: change handling of invalid checksums
If a string is not a sequence of 64 Hex-Characters, then this property
is persisent: checking the same string again will not suddently make it
be a sequence of 64 Hex-Characters. Therefore, make the error persistent.

Also, a syntactically incorrect argument has nothing to do with IO,
so change the underlying exception class.

While there, also improve the error message and include the repository
name as well.

Moreover, as requested in issue #7353, do not abort early, but instead
still download from the specified URL, which will show the hash of the
file at the requested URL.

Related to #7353.

Change-Id: I1126ee55567f9b0753940423e8c80b811190fc8d
PiperOrigin-RevId: 233918098