Release reference to the delegate supplier in InterruptibleSupplier.Memoize.
I don't have any evidence that this is causing a significant memory leak, but it theoretically could be if the delegate supplier holds a reference to an expensive object (which is easy to do unintentionally with a lambda).

Also got rid of unnecessary InterruptibleSupplier.Instance class and added unit tests.

RELNOTES: None.
PiperOrigin-RevId: 234166340