Alter `AuthHeaderRequest` to allow for read only access to HTTP details
!7199 covered changing the method for handling authentication header
creation, moving away from direct dependencies on various downstream
providers to allowing for a more generic interface and contract.

This has been a positive change for the maintainability of the software,
but has, unfortunately had a rubber meets the road problem in attempting
to alter AWS v4-sig authentication headers for the new scheme.

This is down to the fact that the only detail provided for the
authentication scheme is the outgoing URL. This is insufficient for AWS
signature generation which needs to account for the HTTP Method, as well
as some of the headers involved.

As such, we move here to changing the interface slightly, allowing for
read only access to the method and the request HTTP headers. In keeping
with the spirit of !7199, and to allow the flexibility to later change
or alter the implementation details of the remote module, we abstract
this information behind a new `AuthHeaderRequest` interface.

This interface will be further used to implement the logic needed for
!6810.